+++ cart/src/endpoints/cart/router.js
@@ -1,13 +1,13 @@
 /**
  * This class defines all the custom routes in addition to the base routes
- * for `shopping_cart` resource
+ * for `cart` resource
  */
 
 const BaseRouter = require('../base-router');
 
 const Handlers = require('./handlers');
 
-const resource = require('../../constants/modelNames').shopping_cart;
+const resource = require('../../constants/modelNames').cart;
 
 class Router extends BaseRouter {
   constructor(handler) {
@@ -17,7 +17,7 @@ class Router extends BaseRouter {
 
   // add all the resource specific routes here
 
-  // route:  /shopping_cart/a/b/c/d
+  // route:  /cart/a/b/c/d
 }
 
 module.exports = Router;
