+++ docker-compose.yml
@@ -49,9 +49,9 @@ services:
       - "${PGADMIN_PORT:-5050}:80"
     restart: unless-stopped
 
-  shopping_cart:
-    build: ./shopping_cart
-    container_name: "shopping_cart"
+  cart:
+    build: ./cart
+    container_name: "cart"
 
     # Forward ports, format 'container:host'
     ports:
@@ -60,12 +60,12 @@ services:
     # Mount local directories on the container filesystem.
     # Should not happen in production.
     volumes:
-      - ./shopping_cart/src:/usr/app/src
-      - ./shopping_cart/app.js:/usr/app/app.js
-      - ./shopping_cart/process.json:/usr/app/process.json
-      - ./shopping_cart/registrations.js:/usr/app/registrations.js
-      - ./shopping_cart/.shopping_cartrc:/usr/app/.shopping_cartrc
-      - ./shopping_cart/log:/usr/app/log
+      - ./cart/src:/usr/app/src
+      - ./cart/app.js:/usr/app/app.js
+      - ./cart/process.json:/usr/app/process.json
+      - ./cart/registrations.js:/usr/app/registrations.js
+      - ./cart/.cartrc:/usr/app/.cartrc
+      - ./cart/log:/usr/app/log
       # The above line takes care of making the log dir accessible, but should be
       #   handled differently in production.
